# Avoid loading CUDA module at each kernel launch