--- a/runtime/include/chpl-gpu-impl.h
+++ b/runtime/include/chpl-gpu-impl.h
@@ -29,12 +29,12 @@ void chpl_gpu_impl_init(void);
 void chpl_gpu_impl_on_std_modules_finished_initializing(void);
 
 void chpl_gpu_impl_launch_kernel(int ln, int32_t fn,
-                                 const char* fatbinData, const char* name,
+                                 const char* name,
                                  int grd_dim_x, int grd_dim_y, int grd_dim_z,
                                  int blk_dim_x, int blk_dim_y, int blk_dim_z,
                                  int nargs, va_list args);
 void chpl_gpu_impl_launch_kernel_flat(int ln, int32_t fn,
-                                 const char* fatbinPath, const char* name,
+                                 const char* name,
                                  int num_threads, int blk_dim,
                                  int nargs, va_list args);