--- a/runtime/include/chpl-gpu.h
+++ b/runtime/include/chpl-gpu.h
@@ -46,6 +46,32 @@ static inline void CHPL_GPU_DEBUG(const char *str, ...) {
   }
 }
 
+#ifdef CHPL_GPU_ENABLE_PROFILE
+static inline long double get_time() {
+  struct timeval tv;
+
+  gettimeofday(&tv, NULL);
+
+  return (tv.tv_sec*1000000.0+tv.tv_usec)/1000.0;
+}
+
+#define CHPL_GPU_START_TIMER(t) \
+  long double t; \
+  long double start_##t; \
+  do { start_##t = get_time(); } while(0);
+
+#define CHPL_GPU_STOP_TIMER(t) \
+  do { t = get_time()-start_##t; } while(0);
+
+#define CHPL_GPU_PRINT_TIMERS(...) \
+  do { printf(__VA_ARGS__); } while(0);
+
+#else
+#define CHPL_GPU_START_TIMER(t) do {} while(0);
+#define CHPL_GPU_STOP_TIMER(t)
+#define CHPL_GPU_PRINT_TIMERS(...) do {} while(0);
+#endif
+
 static inline bool chpl_gpu_running_on_gpu_locale(void) {
   return chpl_task_getRequestedSubloc()>=0;
 }