--- a/runtime/include/chpl-gpu.h
+++ b/runtime/include/chpl-gpu.h
@@ -82,12 +82,12 @@ void chpl_gpu_on_std_modules_finished_initializing(void);
 void chpl_gpu_get_device_count(int* into);
 
 void chpl_gpu_launch_kernel(int ln, int32_t fn,
-                            const char* fatbinData, const char* name,
+                            const char* name,
                             int grd_dim_x, int grd_dim_y, int grd_dim_z,
                             int blk_dim_x, int blk_dim_y, int blk_dim_z,
                             int nargs, ...);
 void chpl_gpu_launch_kernel_flat(int ln, int32_t fn,
-                                 const char* fatbinPath, const char* name,
+                                 const char* name,
                                  int num_threads, int blk_dim,
                                  int nargs, ...);