--- a/runtime/src/chpl-gpu.c
+++ b/runtime/src/chpl-gpu.c
@@ -43,7 +43,7 @@ void chpl_gpu_on_std_modules_finished_initializing(void) {
 }
 
 inline void chpl_gpu_launch_kernel(int ln, int32_t fn,
-                                   const char* fatbinData, const char* name,
+                                   const char* name,
                                    int grd_dim_x, int grd_dim_y, int grd_dim_z,
                                    int blk_dim_x, int blk_dim_y, int blk_dim_z,
                                    int nargs, ...) {
@@ -62,7 +62,7 @@ inline void chpl_gpu_launch_kernel(int ln, int32_t fn,
   chpl_gpu_diags_incr(kernel_launch);
 
   chpl_gpu_impl_launch_kernel(ln, fn,
-                              fatbinData, name,
+                              name,
                               grd_dim_x, grd_dim_y, grd_dim_z,
                               blk_dim_x, blk_dim_y, blk_dim_z,
                               nargs, args);
@@ -75,7 +75,7 @@ inline void chpl_gpu_launch_kernel(int ln, int32_t fn,
 }
 
 inline void chpl_gpu_launch_kernel_flat(int ln, int32_t fn,
-                                        const char* fatbinData, const char* name,
+                                        const char* name,
                                         int num_threads, int blk_dim, int nargs,
                                         ...) {
 
@@ -94,7 +94,7 @@ inline void chpl_gpu_launch_kernel_flat(int ln, int32_t fn,
   chpl_gpu_diags_incr(kernel_launch);
 
   chpl_gpu_impl_launch_kernel_flat(ln, fn,
-                                   fatbinData, name,
+                                   name,
                                    num_threads, blk_dim,
                                    nargs, args);
   va_end(args);