--- a/runtime/src/gpu/common/cuda-shared.h
+++ b/runtime/src/gpu/common/cuda-shared.h
@@ -28,18 +28,24 @@
 #include "cuda-utils.h"
 
 static inline
-void* chpl_gpu_getKernel(const char* fatbinData, const char* kernelName) {
+void* chpl_gpu_load_module(const char* fatbin_data) {
+  CUmodule cuda_module;
 
-  CUmodule    cudaModule;
-  CUfunction  function;
+  CUDA_CALL(cuModuleLoadData(&cuda_module, fatbin_data));
+  assert(cuda_module);
 
-  // Create module for object
-  CUDA_CALL(cuModuleLoadData(&cudaModule, fatbinData));
+  return (void*)cuda_module;
+}
 
-  // Get kernel function
-  CUDA_CALL(cuModuleGetFunction(&function, cudaModule, kernelName));
+static inline
+void* chpl_gpu_load_function(CUmodule cuda_module, const char* kernel_name) {
+  CUfunction function;
+
+  CUDA_CALL(cuModuleGetFunction(&function, cuda_module, kernel_name));
+  assert(function);
 
   return (void*)function;
+
 }
 
 // this is part of the interface (used by the module code as an extern)