--- a/runtime/src/gpu/cuda/gpu-cuda.c
+++ b/runtime/src/gpu/cuda/gpu-cuda.c
@@ -312,7 +312,7 @@ static void chpl_gpu_launch_kernel_help(int ln,
 #endif
 }
 
-inline void chpl_gpu_impl_launch_kernel(int ln, int32_t fn,
+void chpl_gpu_impl_launch_kernel(int ln, int32_t fn,
                                  const char* fatbinData, const char* name,
                                  int grd_dim_x, int grd_dim_y, int grd_dim_z,
                                  int blk_dim_x, int blk_dim_y, int blk_dim_z,
@@ -324,7 +324,7 @@ inline void chpl_gpu_impl_launch_kernel(int ln, int32_t fn,
                               nargs, args);
 }
 
-inline void chpl_gpu_impl_launch_kernel_flat(int ln, int32_t fn,
+void chpl_gpu_impl_launch_kernel_flat(int ln, int32_t fn,
                                       const char* fatbinData, const char* name,
                                       int num_threads, int blk_dim, int nargs,
                                       va_list args) {