--- a/runtime/src/gpu/cuda/gpu-cuda.c
+++ b/runtime/src/gpu/cuda/gpu-cuda.c
@@ -313,10 +313,15 @@ static void chpl_gpu_launch_kernel_help(int ln,
 }
 
 inline void chpl_gpu_impl_launch_kernel(int ln, int32_t fn,
-                                 const char* fatbinData, const char* name,
-                                 int grd_dim_x, int grd_dim_y, int grd_dim_z,
-                                 int blk_dim_x, int blk_dim_y, int blk_dim_z,
-                                 int nargs, va_list args) {
+                                        const char* fatbinData,
+                                        const char* name,
+                                        int grd_dim_x,
+                                        int grd_dim_y,
+                                        int grd_dim_z,
+                                        int blk_dim_x,
+                                        int blk_dim_y,
+                                        int blk_dim_z,
+                                        int nargs, va_list args) {
   chpl_gpu_launch_kernel_help(ln, fn,
                               fatbinData, name,
                               grd_dim_x, grd_dim_y, grd_dim_z,
@@ -325,9 +330,12 @@ inline void chpl_gpu_impl_launch_kernel(int ln, int32_t fn,
 }
 
 inline void chpl_gpu_impl_launch_kernel_flat(int ln, int32_t fn,
-                                      const char* fatbinData, const char* name,
-                                      int num_threads, int blk_dim, int nargs,
-                                      va_list args) {
+                                             const char* fatbinData,
+                                             const char* name,
+                                             int num_threads,
+                                             int blk_dim,
+                                             int nargs,
+                                             va_list args) {
   int grd_dim = (num_threads+blk_dim-1)/blk_dim;
 
   chpl_gpu_launch_kernel_help(ln, fn,