… runtime/src/gpu/rocm/gpu-rocm.c
… test/gpu/native/streamPrototype/stream.chpl