This PR improves GPU kernel launch times by loading the binary CUDA module only
once per locale.

Previously, we used to re-load the module everytime we launch a kernel. There
was nothing smart behind it -- it was an artifact from earlier prototypes.

The change is relatively straightforward. The only interesting note is that
it appears that you have to load the module per context (i.e., per device).
Which is a bit strange, but that's what the documentation says and not doing
it that way breaks multi-gpu execution.

While there, this test also changes the stream prototype to use `foreach`
instead of `forall` to make it a more apples-to-apples comparison to a
single-GPU C+CUDA reference. `forall` option is available with a `config param`.

On my A100 based system, I am seeing 250 GB/s as opposed to 235 GB/s bandwidth
using 1GB vectors. 250 GB/s is also what I get on an NVIDIA bandwidth test.

Test:
- [x] gpu/native
- [x] AMD prototype