@e-kayrakli 2023-01-09
More performance:

| (M elems/vec)  | pr (GB/s)| main (GB/s)| cuda (GB/s)
|----------------|----------|-----------|---
| 1              | 150.694  | 21.3451   |223.4182
| 2              | 192.106  | 39.1381   |233.2242
| 4              | 210.593  | 65.3234   |237.8802
| 8              | 220.995  | 100.113   |240.4256
| 16             | 219.43   | 134.983   |241.7559
| 32             | 229.563  | 175.831   |242.4080
| 64             | 241.471  | 215.697   |242.3986
| 128            | 250.134  | 235.315   |242.4827

![gpu-stream](https://user-images.githubusercontent.com/4141670/211405042-48c348e2-eb1d-4442-8db2-da6630e89468.png)
![gpu-stream-rel](https://user-images.githubusercontent.com/4141670/211405041-d5e16a95-fe86-4e35-a6ac-60a08a69774d.png)
![gpu-stream-imp](https://user-images.githubusercontent.com/4141670/211405035-ee05f149-9ebe-4d73-b9b0-9a38181192ea.png)

I can't explain how we can beat the C+CUDA version, yet. But I get the same result in multiple tries (tests also repeatedly run stream). The C+CUDA version is not doing anything complicated (neither do we) and it is here: https://github.com/bcumming/cuda-stream.